requirements.txt
torch
torchvision
numpy
Pillow
scikit-learn
tqdm
transformers